// src/interfaces/asset.ts
export interface Asset {
  nid: string;
  assetCreator: string;
  creatorWallet: string;
  assetTimestampCreated: string;
  digitalSourceType: string;
  geolocation: string;
  mimetype: string;
  initialCommitter: string;
  generatedBy: string;
  initialMinter: string;
  license: string;
  miningPreference: string;
  integrityProof: string;
  inStock: number;
}